Bug: None Fix build error

+++ ZimbraMigrationTools/src/c/Exchange/MapiAccessWrap.cpp
@@ -705,10 +705,10 @@ STDMETHODIMP CMapiAccessWrap::GetData(BSTR UserId, VARIANT ItemId, FolderType ty
                                 pIt[attrs[14]] = SysAllocString((apptData.vExceptions[i]->GetPlainTextFileAndContent()).c_str());
                                 pIt[attrs[15]] = SysAllocString(L"text/html");
                                 pIt[attrs[16]] = SysAllocString((apptData.vExceptions[i]->GetHtmlFileAndContent()).c_str());
-                                int numAttendeesInException = apptData.vExceptions[i]->GetAttendees().size();
+                                size_t numAttendeesInException = apptData.vExceptions[i]->GetAttendees().size();
                                 if (numAttendeesInException > 0)
                                 {
-                                    for (int j = 0; j < numAttendeesInException; j++)
+                                    for (size_t j = 0; j < numAttendeesInException; j++)
                                     {                                       
                                         attendeeData += apptData.vExceptions[i]->GetAttendees()[j]->nam;
                                         attendeeData += L"~";
